refactor: Separated Unnecessary Methods into Utility Functions
