+++ src/components/Grids/DateGrid.tsx
@@ -5,10 +5,16 @@ type DatesGridProps = {
   prevMonthDays: number[];
   currentMonthDays: number[];
   nextMonthDays: number[];
-  isToday: (date: number) => boolean;
-  isSelected: (date: number, monthOffset: number) => boolean;
+  isToday: (date: number, currentDate: Date) => boolean;
+  isSelected: (
+    date: number, 
+    monthOffset: number, 
+    currentDate: Date, 
+    selectedDate: Date
+  ) => boolean;
   handleDateSelect: (year: number, month: number, day: number) => void;
   currentDate: Date;
+  selectedDate: Date;
 };
 
 const DateGrid: React.FC<DatesGridProps> = ({
@@ -19,12 +25,13 @@ const DateGrid: React.FC<DatesGridProps> = ({
   isSelected,
   handleDateSelect,
   currentDate,
+  selectedDate,
 }) => (
   <StyledDateGrid>
     {prevMonthDays.map((date) => (
       <ExtraDateUnit
         key={`prev-${date}`}
-        className={isSelected(date, -1) ? 'selected' : ''}
+        className={isSelected(date, -1, currentDate, selectedDate) ? 'selected' : ''}
         onClick={() =>
           handleDateSelect(
             currentDate.getFullYear(),
@@ -36,8 +43,8 @@ const DateGrid: React.FC<DatesGridProps> = ({
       />
     ))}
     {currentMonthDays.map((date) => {
-      const todayClass = isToday(date) ? 'today' : '';
-      const selectedClass = isSelected(date, 0) ? 'selected' : '';
+      const todayClass = isToday(date, currentDate) ? 'today' : '';
+      const selectedClass = isSelected(date, 0, currentDate, selectedDate) ? 'selected' : '';
       const className = selectedClass ? selectedClass : todayClass;
       return (
         <DateUnit
@@ -57,7 +64,7 @@ const DateGrid: React.FC<DatesGridProps> = ({
     {nextMonthDays.map((date) => (
       <ExtraDateUnit
         key={`next-${date}`}
-        className={isSelected(date, 1) ? 'selected' : ''}
+        className={isSelected(date, -1, currentDate, selectedDate) ? 'selected' : ''}
         onClick={() =>
           handleDateSelect(
             currentDate.getFullYear(),
